bump substrate to polkadot-v0.9.2. + add GA + re-add branch ids (#116)
* bump https://github.com/paritytech/substrate/commit/1d7f6e12c651d776fc0dc1adefd007bb60f60b63. Closes #117.
* re-introduce branch identifiers
* Add GitHub actions

#### Rationale Branch Ids:

Polkadot references substrate dependencies including `branch = "master"`. Hence, any rust crate that uses substrate AND Polkadot deps must use the branch-id because Cargo can't identify them as the same even if the same substrate revision is referenced.

This is the api-client branch that is currently used with the encointer-client.

+++ src/examples/example_generic_extrinsic.rs
@@ -41,7 +41,7 @@ fn main() {
         "Balances",
         "transfer",
         GenericAddress::Id(to),
-        Compact(42 as u128)
+        Compact(42_u128)
     );
 
     println!("[+] Composed Extrinsic:\n {:?}\n", xt);
